Fix level detection
