+++ buildSrc/src/main/kotlin/io/github/nucleuspowered/gradle/task/StdOutExec.kt
@@ -1,3 +1,7 @@
+/*
+ * This file is part of Nucleus, licensed under the MIT License (MIT). See the LICENSE.txt file
+ * at the root of this project for more details.
+ */
 package io.github.nucleuspowered.gradle.task
 
 import org.gradle.api.tasks.AbstractExecTask
